[director] Automatically handle sync notes
Note class is modified to work with 3+ notes in sync.

The apparent sync note "pair" comes from the limitation that the game
should be playable with two fingers, not from its intrinsic nature.

It would be a better match for the nature of sync relationship when
unlimited number of notes are allowed to be in sync, and it frees us from
issues arising from the situation in whith users wish to temporarily keep
3+ notes in sync when editing.

TODO: Project save format change